feat and refactor: move icons closer to their feature, as opposed to keeping them in common.
