+++ src/components/About/icons/Values/Group/index.tsx
