+++ src/components/About/icons/Values/Trophy/index.tsx
